pyproject.toml
[project]
name = "omvarldsbevakning-system"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "django",
    "djangorestframework",
    "dj-rest-auth",
    "black",
    "django-filter>=25.1",
    "django-cors-headers>=4.7.0",
]

[tool.uv]

[dependency-groups]
dev = [
    "ruff>=0.12.11",
]

[tool.ruff.lint]
# Ignore the specific comma rule that conflicts with the formatter
ignore = ["COM812"]